requirements.txt
flask>=3.1.1
easyocr>=1.7.2
google-generativeai>=0.8.5
mysql-connector-python>=9.3.0